Release/0.11

+++ bin/build/install-bin-build.sh
@@ -158,7 +158,7 @@ _installBinBuildDirectory() {
   local start tarArgs
   local target="$applicationHome/build.tar.gz"
 
-  if [ -z "$localPath" ]; then
+  if [ -n "$localPath" ]; then
     _installBinBuildDirectoryLocal "$usage" "$applicationHome" "$localPath"
     return $?
   fi
@@ -182,9 +182,13 @@ _installBinBuildDirectoryLocal() {
   # Clean target regardless
   backupPath="$installPath.aboutToDelete.$$"
   __usageEnvironment "$usage" rm -rf "$backupPath" || return $?
-  __usageEnvironment "$usage" mv -f "$installPath" "$backupPath" || return $?
-  __usageEnvironment "$usage" cp -r "$localPath" "$installPath" || _undo $? rf -f "$installPath" || _undo $? mv -f "$backupPath" "$installPath" || return $?
-  __usageEnvironment "$usage" rm -rf "$backupPath" || :
+  if [ -d "$installPath" ]; then
+    __usageEnvironment "$usage" mv -f "$installPath" "$backupPath" || return $?
+    __usageEnvironment "$usage" cp -r "$localPath" "$installPath" || _undo $? rf -f "$installPath" || _undo $? mv -f "$backupPath" "$installPath" || return $?
+    __usageEnvironment "$usage" rm -rf "$backupPath" || :
+  else
+    __usageEnvironment "$usage" cp -r "$localPath" "$installPath" || return $?
+  fi
 }
 
 # Check the build directory after installation
